requirements.txt
streamlit==1.27.2
PyPDF2==3.0.1
pdfplumber==0.10.2
Pillow==10.0.1
pdf2image==1.16.3
reportlab==4.0.4